+++ CHANGELOG.rst
@@ -11,14 +11,20 @@ New Features:
 * Now copies over data from the document catalog to the cropped document.
   This includes, for example, the outline or bookmarks.
 
-* There is a new option `--docCatBlacklist` (shortcut `-dcb`) which can
+* There is a new option ``--docCatBlacklist`` (shortcut ``-dcb``) which can
   be used to block any particular item from being copied.  The default is
   an empty string, which copies everything possible.  To revert to the
-  previous behavior of pdfCropMargins you can set `-dcb "ALL"`.
+  previous behavior of pdfCropMargins you can set ``-dcb "ALL"``.  See
+  the program's help option ``-h``.
 
-* There is another new option `--docCatWhitelist` (shortcut `dcw`) which
+* There is another new option ``--docCatWhitelist`` (shortcut ``dcw``) which
   is a list of document catalog items to always try to copy over.  This
   list overrides the blacklist.
+
+* There is a new option to use an order statistic in choosing the page size for
+  the ``--samePageSize`` option.  The argument is the number ``n`` of pages to
+  ignore in each edge calculation.  The option is ``--samePageSizeOrderStat``
+  (shortcut ``-ms``).  See the program's help option ``-h``.
 
 0.1.2 (2017-03-14)
 ------------------